upd: create modal form component for all assets transacions, changed logic according to a new flow
