+++ src/components/layout/LayoutPortfolio/Sidebar/Sidebar.tsx
@@ -1,11 +1,11 @@
 'use client'
 
 import Image from 'next/image'
-import { SidebarMenu } from '@/components/layout/LayoutPortfolio/Sidebar/SidebarMenu/SidebarMenu'
 import Link from 'next/link'
+import { useRef, useState } from 'react'
+import { SidebarMenu } from '@/components/layout/LayoutPortfolio/Sidebar/SidebarMenu/SidebarMenu'
 import { PUBLIC_PAGE } from '@/config/paths.config'
 import { ChevronLeft, ChevronRight } from 'lucide-react'
-import { useRef, useState } from 'react'
 
 export function Sidebar() {
     const [collapsedMenu, setCollapsedMenu] = useState<boolean>(false)
