+++ src/components/layout/LayoutPortfolio/Sidebar/SidebarMenu/SidebarMenu.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-// import Link from 'next/link'
 import { SIDEBAR_DATA } from './sidebarMenu.data'
 import type { ISidebarItem } from './sidebarMenu.types'
 import { useEffect, useRef, useState } from 'react'
