--- src/components/ui/Modal/modal.types.ts
@@ -1,6 +1,0 @@
-export interface ModalProps {
-    isOpen: boolean
-    onClose: () => void
-    modalTitle?: string
-    children?: React.ReactNode
-}
